# backend/docs/swagger.yaml
basePath: /api
definitions:
  handler.UpdateAvatarRequest:
    properties:
      avatar_url:
        example: /assets/default_avatars/cat.jpg
        type: string
    required:
    - avatar_url
    type: object
  model.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.AuthResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/model.AuthResponse'
    type: object
  model.CSVImportRequest:
    properties:
      content:
        example: |-
          jp,ru
          猫,кот
        type: string
    required:
    - content
    type: object
  model.CSVImportResponse:
    properties:
      errors:
        example:
        - 'строка 1: неверный формат'
        items:
          type: string
        type: array
      failed_count:
        example: 0
        type: integer
      imported_count:
        example: 10
        type: integer
    type: object
  model.CSVImportResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/model.CSVImportResponse'
    type: object
  model.ErrorResponse:
    properties:
      error:
        example: неверные данные
        type: string
    type: object
  model.Role:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
  model.User:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.Role'
        example: user
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  model.UserLoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secretpassword
        type: string
    required:
    - email
    - password
    type: object
  model.UserRegisterRequest:
    properties:
      email:
        example: newuser@example.com
        type: string
      name:
        example: New User
        type: string
      password:
        example: secretpassword
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  model.UserResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/model.User'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для личного русско-японского словаря
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: JP-RU Dictionary API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает JWT токен.
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/model.AuthResponseWrapper'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Вход пользователя
      tags:
      - auth
  /api/auth/logout:
    post:
      description: Валидирует выход пользователя (на клиенте удаляется токен)
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Выход пользователя
      tags:
      - auth
  /api/auth/me:
    get:
      description: Возвращает информацию о пользователе из JWT токена. Требует Bearer
        токен.
      produces:
      - application/json
      responses:
        "200":
          description: Информация о пользователе
          schema:
            $ref: '#/definitions/model.UserResponseWrapper'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Информация о текущем пользователе
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает JWT токен. Требует email,
        пароль и имя.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/model.AuthResponseWrapper'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
...
  /api/users/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Загружает аватар пользователя (JPG/PNG) и обновляет профиль.
      parameters:
      - description: Файл аватара
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: URL аватара
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка загрузки или неверный формат
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузка аватара
      tags:
      - users
  /api/users/avatar/url:
    put:
      consumes:
      - application/json
      description: Устанавливает аватар пользователя по указанному URL (обычно для
        предустановленных аватаров).
      parameters:
      - description: URL аватара
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успех
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление аватара по URL
      tags:
      - users
...
  /api/words/import:
    post:
      consumes:
      - application/json
      description: 'Импортирует слова из CSV формата. Формат: jp;ru;on;kun;ex_jp;ex_ru;tags.
        Поля разделены точкой с запятой, массивы внутри полей разделены запятой.'
      parameters:
      - description: CSV данные для импорта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CSVImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Результат импорта
          schema:
            $ref: '#/definitions/model.CSVImportResponseWrapper'
        "400":
          description: Неверный формат CSV
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Импорт слов из CSV
      tags:
      - words
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"